pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "debriev"
version = "0.1.0"
description = "Backend-first litigation drafting and verification engine."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.14.0",
  "fastapi>=0.115.0",
  "psycopg[binary]>=3.2.0",
  "pydantic-settings>=2.6.0",
  "sqlalchemy>=2.0.36",
  "uvicorn>=0.32.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.0",
  "pytest>=8.3.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["app/tests"]
pythonpath = ["."]

